add document service and readjust the flow

+++ DataAccess/Repository.cs
@@ -1,6 +1,6 @@
-using Entities;
+using EnigmaAPI.Entities;
 
-namespace DataAccess;
+namespace EnigmaAPI.DataAccess;
 
 public class Repository<T> : IRepository<T> where T : IEntity
 {
@@ -16,9 +16,12 @@ public class Repository<T> : IRepository<T> where T : IEntity
         return _dataStorage.GetAllAsync();
     }
 
-    public Task<T> GetByIdAsync(int id)
+    public Task<T> GetByIdAsync(string id)
     {
-        return _dataStorage.GetAsync(d => d.Id == id);
+        if (string.IsNullOrEmpty(id))
+            throw new Exception($"{nameof(GetByIdAsync)} failed to execute, a guid should be provided");
+
+        return _dataStorage.GetAsync(d => d.Id == new Guid(id));
     }
 
     public Task<IEnumerable<T>> GetAllWhereAsync(Func<T, bool> predicate)
@@ -41,8 +44,11 @@ public class Repository<T> : IRepository<T> where T : IEntity
         return _dataStorage.UpdateAsync(entity);
     }
 
-    public Task DeleteAsync(int id)
+    public Task DeleteAsync(string id)
     {
+        if (string.IsNullOrEmpty(id))
+            throw new Exception($"{nameof(GetByIdAsync)} failed to execute, a guid should be provided");
+
         return _dataStorage.DeleteAsync(id);
     }
 }
